updated website content with real information
