+++ src/components/Landing.tsx
@@ -1,37 +1,36 @@
 import React from "react";
 import Navbar from "./Navbar";
 import loginImg from "./assets/login_4.png";
-import LandCopy from "./LandCopy";
-import SplashInfoLeft from "./SplashInfoLeft";
-import SplashInfoRight from "./SplashInfoRight";
-import SplashTeam from './SplashTeam';
-import GetKlusterInfo from "./GetKlusterInfo";
-import { Login } from "./Login";
+import LandCopy from "./LandingDownloadLink";
+import SplashInfoLeft from "./LandingInfoLeft";
+import SplashInfoRight from "./LandingInfoRight";
+import SplashTeam from "./LandingTeam";
+import splash5 from "./assets/splash5.png";
+import splash6 from "./assets/splash6.png";
 
 function Landing() {
-  const navItems = ["Login", "help"].map((item, i) => {
+  const navItems = ["Contact", "About"].map((item, i) => {
     return <li key={item + i}>{item}</li>;
   });
 
   const infoItems = [
     {
-      "header": "Save money with Kuberenetes!",
-      "text": "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Iste nemo atque veniam aut! Itaque architecto nihil repudiandae minus perspiciatis, dolores quaerat ullam, recusandae beatae et quasi a optio illum veniam!",
+      header: "Track your Resource Utilization with Kubehawk's Dashboard",
+      text: "Gain a competitive edge by closely monitoring your Kubernetes resource usage. Our powerful visualizer offers detailed insights into the consumption of CPU, memory, and storage, enabling you to optimize deployments and maximize the efficiency of your infrastructure.",
     },
     {
-      "header": "Save money with Kuberenetes!",
-      "text": "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Iste nemo atque veniam aut! Itaque architecto nihil repudiandae minus perspiciatis, dolores quaerat ullam, recusandae beatae et quasi a optio illum veniam!",
+      header: "Visualize your Cluster's RBAC Data in a Simplified Way",
+      text: "Empower your Kubernetes administrators with our cutting-edge visualizer that provides a comprehensive view of Role-Based Access Control information, ensuring optimal security and streamlined management of your cluster permissions.",
     },
     {
-      "header": "Save money with Kuberenetes!",
-      "text": "Lorem ipsum dolor sit, amet consectetur adipisicing elit. Iste nemo atque veniam aut! Itaque architecto nihil repudiandae minus perspiciatis, dolores quaerat ullam, recusandae beatae et quasi a optio illum veniam!",
+      header: "Seamlessly Integrate Kubehawk for Immediate Impact",
+      text: "Deploy our open-source metrics visualizer directly within your existing Kubernetes cluster for a rapid setup and instant value. Enhance your organization's monitoring capabilities and gain critical insights without any additional hassle or expense.",
     },
-
-  ]
+  ];
 
   return (
     <div className="landing">
-      <Navbar navItems={navItems} kubeButton={<button>Get KubeHawk</button>} />
+      <Navbar id="landing-nav" navItems={navItems} kubeButton={<button>Get KubeHawk</button>} />
       <div id="landing-body">
         <div id="landing-login">
           <LandCopy />
@@ -39,9 +38,21 @@ function Landing() {
             <img src={loginImg} alt="security_login" />
           </div>
         </div>
-        <SplashInfoLeft header={infoItems[0].header} text={infoItems[0].text} image={loginImg}></SplashInfoLeft>
-        <SplashInfoRight header={infoItems[1].header} text={infoItems[1].text} image={loginImg}></SplashInfoRight>
-        <SplashInfoLeft header={infoItems[2].header} text={infoItems[2].text} image={loginImg}></SplashInfoLeft>
+        <SplashInfoLeft
+          header={infoItems[0].header}
+          text={infoItems[0].text}
+          image={splash5}
+        ></SplashInfoLeft>
+        <SplashInfoRight
+          header={infoItems[1].header}
+          text={infoItems[1].text}
+          image={splash6}
+        ></SplashInfoRight>
+        <SplashInfoLeft
+          header={infoItems[2].header}
+          text={infoItems[2].text}
+          image={loginImg}
+        ></SplashInfoLeft>
         <SplashTeam></SplashTeam>
       </div>
     </div>
