--- src/components/Splash.tsx
@@ -1,21 +1,0 @@
-import React from 'react'
-import hawkGIF from './assets/curious.gif'
-
-function Splash() {
-  return (
-    <div id='splash-main'>
-      <div id="splash-title">
-        <h1>KubeHawk</h1>
-        <h2>A Kubernetes Securrrrity Monitor</h2>
-    </div>
-    <div id="enter">
-    <button >Dashboard</button>
-    </div>
-    </div>
-
-  )
-}
-
-
-
-export default Splash
