--- src/components/SplashInfoLeft.tsx
@@ -1,20 +1,0 @@
-import React from "react";
-
-
-const SplashInfoLeft = (props) => {
-    return (
-      <div className="splashleft">
-        <div id="splash.info">
-          <h1 id="splash.header">
-            {props.header}
-          </h1>
-          <div id="splash.text">
-            {props.text}
-          </div>
-        </div>
-        <img src={props.image} className="splashleftimg" alt="A picture" />
-      </div>
-    )
-}
-
-export default SplashInfoLeft;
