--- src/components/SplashInfoRight.tsx
@@ -1,20 +1,0 @@
-import React from "react";
-
-
-const SplashInfoRight = (props) => {
-    return (
-      <div className="splashleft">
-        <img src={props.image} className="splashleftimg" alt="A picture" />
-        <div id="splash.info">
-          <h1 id="splash.header">
-          {props.header}
-          </h1>
-          <div id="splash.text">
-            {props.text}
-          </div>
-        </div>
-      </div>
-    )
-}
-
-export default SplashInfoRight;
